+++ Source/Public/ConvertFrom-Expression.ps1
@@ -3,8 +3,8 @@
     Deserializes a PowerShell expression to an object.
 
 .DESCRIPTION
-    The `ConvertFrom-Expression` cmdlet converts a PowerShell formatted expression to an object-graph existing of
-    a mixture of nested arrays, hashtables and objects that contain a list of strings and values.
+    The `ConvertFrom-Expression` cmdlet safely converts a PowerShell formatted expression to an object-graph
+    existing of a mixture of nested arrays, hashtables and objects that contain a list of strings and values.
 
 .PARAMETER InputObject
     Specifies the PowerShell expressions to convert to objects. Enter a variable that contains the string,
@@ -22,22 +22,35 @@
 
     > [!Caution]
     >
-    > In full language mode the concerned string will simply be invoke using [Invoke-Expression].
+    > In full language mode, `ConvertTo-Expression` permits all type initializers. Cmdlets, functions,
+    > CIM commands, and workflows will *not* be invoked by the `ConvertFrom-Expression` cmdlet.
     >
-    > Take reasonable precautions when using the Invoke-Expression cmdlet in scripts. When using
-    > `-LanguageMode Full` to run a command that the user enters, verify that the command is safe to run
-    > before running it. In general, it is best to design your script with predefined input options,
-    > rather than allowing freeform input.
+    > Take reasonable precautions when using the `Invoke-Expression -LanguageMode Full` command in scripts.
+    > Verify that the class types in the expression are safe before instantiating them. In general, it is
+    > best to design your configuration expressions with restricted or constrained classes, rather than
+    > allowing full freeform expressions.
+
+.PARAMETER ArrayAs
+    If supplied, the array subexpression `@( )` syntaxes without an type initializer or with an unknown or
+    denied type initializer will be converted to the given list type.
+
+.PARAMETER HashTableAs
+    If supplied, the array subexpression `@{ }` syntaxes without an type initializer or with an unknown or
+    denied type initializer will be converted to the given map (dictionary or object) type.
 
 #>
 function ConvertFrom-Expression {
-    [CmdletBinding()][OutputType([Object])] param(
+    [CmdletBinding(HelpUri='https://github.com/iRon7/ObjectGraphTools/blob/main/Docs/ConvertFrom-Expression.md')][OutputType([Object])] param(
 
         [Parameter(Mandatory = $true, ValueFromPipeLine = $True)]
         [Alias('Expression')][String]$InputObject,
 
         [ValidateScript({ $_ -ne 'NoLanguage' })]
-        [PSLanguageMode]$LanguageMode = 'Restricted'
+        [System.Management.Automation.PSLanguageMode]$LanguageMode = 'Restricted',
+
+        [ValidateNotNull()]$ArrayAs,
+
+        [ValidateNotNull()]$HashTableAs
     )
 
     begin {
@@ -48,9 +61,59 @@ function ConvertFrom-Expression {
         }
 
         if ($this.LanguageMode -eq 'NoLanguage') { Throw 'The language mode "NoLanguage" is not supported.' }
+
+        function NewNode($Object) { # Should become a New-Node cmdlet
+            if ($Null -eq $Object) { return }
+            elseif ($Object -is [String] -or $Object -is [Type]) {
+                $String = "$Object" # This will use the type accessor for [Type] type.
+                if ('Array', 'Object[]' -eq $String) {
+                    $Object = @()
+                }
+                elseif ('ordered', 'OrderedDictionary' -eq $String) {
+                    $Object = [System.Collections.Specialized.OrderedDictionary]::new([StringComparer]::OrdinalIgnoreCase)
+                }
+                elseif ('hashtable' -eq $String) {
+                    $Object = [HashTable]::new([StringComparer]::OrdinalIgnoreCase)
+                }
+                elseif ('psobject', 'PSCustomObject' -eq $String) {
+                    $Object = [PSCustomObject]@{}
+                }
+                else {
+                    $Type = $TypeName -as [Type]
+                    if (-not $Type) { StopError "Unknown type: [$Object]" }
+                    $Object = New-Object -Type $TypeName
+                }
+            }
+            elseif ($Null -ne $Object.GetType().GetMethod('Clear')) {
+                $Object.Clear()
+            }
+            [PSNode]::ParseInput($Object)
+        }
+
+        $ArrayNode     = NewNode $ArrayAs
+        $HashTableNode = NewNode $HashTableAs
+
+        if (
+            $ArrayNode -is [PSMapNode] -and $HashTableNode -is [PSListNode] -or
+            -not $ArrayNode -and $HashTableNode -is [PSListNode] -or
+            $ArrayNode -is [PSMapNode] -and -not $HashTableNode
+        ) {
+            $ArrayNode, $HashTableNode = $HashTableNode, $ArrayNode # In case the parameter positions are swapped
+        }
+
+        $ArrayType = if ($ArrayNode) {
+            if ($ArrayType -is [PSListNode]) { $ArrayNode.ValueType }
+            else { StopError 'The -ArrayAs parameter requires a string, type or an object example that supports a list structure' }
+        }
+
+        $HashTableType = if ($HashTableNode) {
+            if ($HashTableNode -is [PSMapNode]) { $HashTableNode.ValueType }
+            else { StopError 'The -HashTableAs parameter requires a string, type or an object example that supports a map structure' }
+        }
+        if ('System.Management.Automation.PSCustomObject' -eq $HashTableNode.ValueType) { $HashTableType = 'PSCustomObject' -as [type] } # https://github.com/PowerShell/PowerShell/issues/2295
     }
 
     process {
-        [PSDeserialize]::new($InputObject, $LanguageMode).Object
+        [PSDeserialize]::new($InputObject, $LanguageMode, $ArrayType, $HashTableType).Object
     }
 }